requirements.txt
asgiref==3.7.2
Django==5.0.3
pillow==10.2.0
sqlparse==0.4.4
django-import-export==3.3.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
django-filter==24.2